Added prototype files for printing and example usage
table.py is a custom data storage and display class.
useful for creatinmg pleasing tables in text to display results.

objgetter.py is an expirimental program which retreives all
modules in the directory "C:\\Users\\username\\AppData\\Local\\Programs\\Python\\Python312\\Lib"
which contains a list of all the users possible imports, excluding
local imports. though if there are other imports added to the
program they are included as well as they are part of the session.

NOTE: currently not functioning. in development. i added them to
enable version control (track changes).